Improve test runner: allow skipping preparation and specify testfiles
Call run_tests.sh with no args to get an help screen

+++ test/run_tests.sh
@@ -4,9 +4,25 @@
 # TODO: read from there
 REDIS_PORT=6333
 
+OPT_CREATE=yes # create the test environment
+OPT_DROP=yes   # drop the test environment
+
+cd $(dirname $0)
+BASEDIR=$(pwd)
+cd -
+
 cleanup() {
-	echo "Cleaning up"
-	kill ${PID_REDIS}
+  if test x"$OPT_DROP" = xyes; then
+    if test x"$PID_REDIS" = x; then
+      PID_REDIS=$(cat ${BASEDIR}/redis.pid)
+      if test x"$PID_REDIS" = x; then
+        echo "Could not find a test redis pid to kill it"
+        return;
+      fi
+    fi
+    echo "Cleaning up"
+    kill ${PID_REDIS}
+  fi
 }
 
 cleanup_and_exit() {
@@ -23,23 +39,43 @@ die() {
 
 trap 'cleanup_and_exit' 1 2 3 5 9 13
 
-echo "Starting redis on port ${REDIS_PORT}"
-echo "port ${REDIS_PORT}" | redis-server - > test/test.log &
-PID_REDIS=$!
+while [ -n "$1" ]; do
+        if test "$1" = "--nodrop"; then
+                OPT_DROP=no
+                shift
+                continue
+        elif test "$1" = "--nocreate"; then
+                OPT_CREATE=no
+                shift
+                continue
+        else
+                break
+        fi
+done
 
-echo "Preparing the environment"
-cd test; sh prepare_db.sh || die "database preparation failure"; cd -;
+if [ -z "$1" ]; then
+        echo "Usage: $0 [<options>] <test> [<test>]" >&2
+        echo "Options:" >&2
+        echo " --nocreate   do not create the test environment on start" >&2
+        echo " --nodrop     do not drop the test environment on exit" >&2
+        exit 1
+fi
+
+TESTS=$@
+
+if test x"$OPT_CREATE" = xyes; then
+  echo "Starting redis on port ${REDIS_PORT}"
+  echo "port ${REDIS_PORT}" | redis-server - > ${BASEDIR}/test.log &
+  PID_REDIS=$!
+  echo ${PID_REDIS} > ${BASEDIR}/redis.pid
+
+  echo "Preparing the environment"
+  cd ${BASEDIR}; sh prepare_db.sh || die "database preparation failure"; cd -
+fi
 
 PATH=node_modules/.bin/:$PATH
 
 echo "Running tests"
-mocha -u tdd \
-  test/unit/redis_pool.test.js \
-  test/unit/metadata.test.js \
-  test/unit/oauth.test.js \
-  test/unit/psql.test.js \
-  test/acceptance/app.test.js  \
-  test/acceptance/app.auth.test.js 
-
+mocha -u tdd ${TESTS}
 
 cleanup
